requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
PyMySQL==1.1.1
Flask-Login==0.6.3
python-dotenv==1.1.1
gunicorn==20.1.0
